refactor: dedicated type file for agenda
